Refactor docker usage
