+++ OrgProbe/amqpqueue.py
@@ -1,7 +1,5 @@
 import json
-
 import logging
-
 import pika
