+++ config/docker.config.ini
@@ -1,0 +1,25 @@
+# Variables enclosed with %() will be interpolated with environment variables.
+
+[global]
+interval = 1
+
+[api]
+https=True
+host=api.blocked.org.uk
+port=443
+version=1.2
+
+[probe]
+uuid=%(PROBE_UUID)
+secret=%(PROBE_SECRET)
+queue=org
+selftest=True
+public_ip=127.0.0.1
+
+[amqp]
+host=api.blocked.org.uk
+port=5672
+ssl=False
+userid=%(AMQP_USER)
+passwd=%(AMQP_PASS)
+prefetch=200
