--- docker/config.ini.tmpl
@@ -1,29 +1,0 @@
-[global]
-interval = 0
-
-[api]
-host=API_HOST
-port=API_PORT
-https=API_HTTPS
-version=1.2
-
-[amqp]
-host=API_HOST
-port=5672
-ssl=True
-userid=AMQP_USER
-passwd=AMQP_PASSWD
-
-
-[public]
-uuid=PROBE_UUID
-secret=PROBE_SECRET
-#heartbeat=240
-selftest = true
-lifetime = 500
-queue = public
-#network = {{network_name}}
-read_size=0
-verify_ssl=false
-
-
