+++ orgprobe
@@ -1,12 +1,10 @@
 #!/usr/bin/env python
 
-import sys
 import argparse
+import os
+import sys
 
-try:
-    from ConfigParser import ConfigParser
-except ImportError:
-    from configparser import ConfigParser
+import configparser
 import logging
 
 from OrgProbe import Probe, APIRequest
@@ -35,8 +33,14 @@ logging.getLogger('urllib3.connectionpool').setLevel(logging.ERROR)
 logging.getLogger('pika').setLevel(logging.ERROR)
 logging.getLogger('requests.packages.urllib3.connectionpool').setLevel(logging.ERROR)
 
-config = ConfigParser()
-loaded = config.read([args.config])
+logging.error(type(os.environ))
+
+config = configparser.ConfigParser(
+    interpolation=configparser.ExtendedInterpolation(),
+    defaults=os.environ)
+
+loaded = config.read(args.config)
+
 logging.info("Loaded %s config files from %s", loaded, args.config)
 
 if config.has_section('api'):
