+++ requirements.txt
@@ -1,5 +1,6 @@
+configparser>=3.5.0
 pika>=0.10.0
 requests==2.12.0
 urllib3==1.21.1
-pyasn1
-redis
+pyasn1>=0.3.7
+redis>=2.10.6
