+++ tox.ini
@@ -1,5 +1,6 @@
 [tox]
 envlist = py27,pep8
+skipsdist = True
 
 [testenv]
 deps=
